spark_job/job/requirements.txt
pyspark==3.5.0
pymongo==4.6.1
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2
transformers==4.36.2
torch==2.1.2
spacy==3.7.2
nltk==3.8.1
python-dotenv==1.0.0